park-finder-api/demo/requirements.txt
gspread
oauth2client
dnspython
pytz
Flask == 2.0.1
Werkzeug==2.2.2
autopep8==2.0.2
flask-cors==3.0.10
python-dotenv==1.0.0
pymongo==3.12.3
pymongo[srv]
